move vulnerablepackagelistitem into pages folder
