+++ lunatrace/bsl/frontend/src/pages/project/builds/VulnerablePackageListItem/VulnerabilityTableItem.tsx
@@ -16,11 +16,11 @@ import classNames from 'classnames';
 import React, { useState } from 'react';
 import { FloatingLabel, Form, FormControl, OverlayTrigger, Spinner, Tooltip } from 'react-bootstrap';
 import { XSquare } from 'react-feather';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 
-import api from '../../api';
-import { ConfirmationDailog } from '../ConfirmationDialog';
-import { CvssInferredWarning } from '../CvssInferredWarning';
+import api from '../../../../api';
+import { ConfirmationDailog } from '../../../../components/ConfirmationDialog';
+import { CvssInferredWarning } from '../../../../components/CvssInferredWarning';
 
 import { Finding } from './types';
